third_party/blink/web_tests/external/wpt/html/interaction/focus/the-autofocus-attribute/supported-elements-expected.txt

This is a testharness.js-based test.
PASS Contenteditable element should support autofocus
PASS Element with tabindex should support autofocus
PASS Non-HTMLElement should not support autofocus
PASS Host element with delegatesFocus should support autofocus
FAIL Host element with delegatesFocus including no focusable descendants should be skipped assert_equals: expected Element node <input autofocus=""></input> but got Element node <body><div autofocus=""></div><input autofocus=""></body>
PASS Area element should support autofocus
Harness: the test ran to completion.
